Move cypher functions & mapping to new query folder
