+++ src/core/database/query/mapping.ts
